requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
djangorestframework==3.13.1
importlib-metadata==4.11.3
Jinja2==3.1.1
Markdown==3.3.6
MarkupSafe==2.1.1
pdfkit==1.0.0
Pillow==9.1.0
pytz==2022.1
qrcode==7.3.1
sqlparse==0.4.2
zipp==3.8.0
